etl/2_History/src/requirements.txt
psycopg2-binary~=2.9.6
python-dotenv~=1.0.1